Better extension management
