+++ Occcam News/Networking/JSONParser.swift
@@ -7,29 +7,6 @@
 
 import Foundation
 
-extension ISO8601DateFormatter {
-    convenience init(_ formatOptions: Options) {
-        self.init()
-        self.formatOptions = formatOptions
-    }
-}
-
-extension Formatter {
-    static let iso8601withFractionalSeconds = ISO8601DateFormatter([.withInternetDateTime, .withFractionalSeconds])
-}
-
-extension JSONDecoder.DateDecodingStrategy {
-    static let iso8601withFractionalSeconds = custom {
-        let container = try $0.singleValueContainer()
-        let string = try container.decode(String.self)
-        guard let date = Formatter.iso8601withFractionalSeconds.date(from: string) else {
-            throw DecodingError.dataCorruptedError(in: container,
-                  debugDescription: "Invalid date: " + string)
-        }
-        return date
-    }
-}
-
 class JSONParser: Parser {
     
     let decoder: JSONDecoder
